rm external/pcmsolver submodule reference
On master this will be fetched at build time
and not referenced via submodules.
This manual modification is done on Dalton2016
because Dalton2016 and master have diverged
too much and because external/pcmsolver is
not shipped with the release tarball.

This change will be merged to master
with -s ours as an empty merge to not
confuse the history even more.