hjaaj: use PARHER for parallel 2-electron integral calculation
instead of PARCAL. PARCAL is now truly the global variable telling
that this is MPI parallel, PARHER tells specifically to CALL PARDRV
in Hermit for parallel 2-electron integral calculation.
PARHER is of course .false. if PARCAL is .false.
However, now we can set PARHER false if we want to parallelize
on a higher level than in Hermit. Used in .FCKTRA.
Could also be used for e.g. parallel numerical differentiation.