hjaaj: new option .VIRTRUNC for deleting diffuse virtual/secondary orbitals
Input:

.VIRTRUNC
 n thr

n   : 2 or 4, using r**2 or r**4 for property matrix
thr : virtual MO eigenvectors to property matrix with eigenvalue .gt. thr**n
      are removed

The truncated MOs are written to SIRIUS.RST and Dalton stops.
It is too complicated for now to update SIRIFC with new set of orbitals, therefore the stop.
Start Dalton again with .MOSTART NEWORB.