Changed the cc_energy_triplet tests
These no-longer check for the printed % of singles contribution
to the transitions. These values are not well-defined anyhow, as
they depend entirely on the parametrization of the excitation operators.

Also the cm-1 value of one excitation on cc_energy_triplet2 to has
been relaxed (but only in cm-1, so the overall test should be equally reliable).