setting up tests for importing static embedding potential
set up two tests for the reading of a precalculated embedding potential
and construction of the corresponding contributions to the 1-electron
fock matrix. with the resulting orbitals expectation values (dipole etc)
are calculated with different wavefunctions.

the first test is done with a low-accuracy potential and a small basis,
and another with avtz bases and a more accurate potential.

the systems is a water molecule surrounded by two water molecules.