BigFix: Remove uninitialized variable 'SHF' -- thanks to HJJ for spotting this.
Also added check to prevent run for restricted open-shell DFT. The AC relies on
a connection between the HOMO eigenvalue and -I. For RO-KS some thought would be
required to develop a suitable correction... so to address issue #72 I suggest
we simply quit if this is invoked with RO-KS.

The procedure is setup to work with UKS, for energies / SCF, hence input alpha/beta
values but this is not present in the main branch AFAIK, and would not carry over to
properties etc. automatically anyway.

AMT